feat: repo's `FindByUsernameOrEmail` returns the error
