+++ src/repo/repo.go
@@ -8,7 +8,7 @@ import (
 
 type Repository interface {
 	Create(user *domain.User) error
-	FindByUsernameOrEmail(username, email string) *domain.User // TODO: Make this return (user, error)
+	FindByUsernameOrEmail(username, email string) (*domain.User, error)
 	FindById(id uint) (*domain.User, error)
 	Update(user *domain.User) error
 }
@@ -25,13 +25,13 @@ func (r *repository) Create(user *domain.User) error {
 	return r.db.Create(user).Error
 }
 
-func (r *repository) FindByUsernameOrEmail(username, email string) *domain.User {
+func (r *repository) FindByUsernameOrEmail(username, email string) (*domain.User, error) {
 	var user domain.User
 	err := r.db.Where("username = ? OR email = ?", username, email).First(&user).Error
 	if err != nil {
-		return nil
+		return nil, err
 	}
-	return &user
+	return &user, nil
 }
 
 func (r *repository) FindById(id uint) (*domain.User, error) {
